VO 주석 추가

+++ src/main/java/com/rocket/rocket/domain/ScoreVO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ScoreVO {
 
-	private int lec_sc_star;
+	private double lec_sc_star;
 }
